@@ -3,10 +3,9 @@
 // found in the LICENSE file.
 
 #import "FLTVideoPlayerPlugin.h"
-
 #import <AVFoundation/AVFoundation.h>
 #import <GLKit/GLKit.h>
-
+#import <AVKit/AVKit.h>
 #import "AVAssetTrackUtils.h"
 #import "messages.g.h"
 
@@ -33,7 +32,7 @@
 }
 @end
 
-@interface FLTVideoPlayer : NSObject <FlutterTexture, FlutterStreamHandler>
+@interface FLTVideoPlayer : NSObject <FlutterTexture, FlutterStreamHandler, AVPictureInPictureControllerDelegate>
 @property(readonly, nonatomic) AVPlayer *player;
 @property(readonly, nonatomic) AVPlayerItemVideoOutput *videoOutput;
 @property(readonly, nonatomic) CADisplayLink *displayLink;
@@ -44,6 +43,8 @@
 @property(nonatomic, readonly) BOOL isPlaying;
 @property(nonatomic) BOOL isLooping;
 @property(nonatomic, readonly) BOOL isInitialized;
+@property(nonatomic) AVPlayerLayer* _playerLayer;
+@property(nonatomic) bool _pictureInPicture;
 - (instancetype)initWithURL:(NSURL *)url
                frameUpdater:(FLTFrameUpdater *)frameUpdater
                 httpHeaders:(nonnull NSDictionary<NSString *, NSString *> *)headers;
@@ -56,6 +57,12 @@ static void *durationContext = &durationContext;
 static void *playbackLikelyToKeepUpContext = &playbackLikelyToKeepUpContext;
 static void *playbackBufferEmptyContext = &playbackBufferEmptyContext;
 static void *playbackBufferFullContext = &playbackBufferFullContext;
+
+#if TARGET_OS_IOS
+API_AVAILABLE(ios(9.0))
+AVPictureInPictureController *_pipController;
+#endif
+
 
 @implementation FLTVideoPlayer
 - (instancetype)initWithAsset:(NSString *)asset frameUpdater:(FLTFrameUpdater *)frameUpdater {
@@ -232,8 +239,63 @@ NS_INLINE CGFloat radiansToDegrees(CGFloat radians) {
   [self addObservers:item];
 
   [asset loadValuesAsynchronouslyForKeys:@[ @"tracks" ] completionHandler:assetCompletionHandler];
+  [self usePlayerLayer];
 
   return self;
+}
+- (void)setPictureInPicture:(BOOL)pictureInPicture {
+#if TARGET_OS_IOS
+    if (self._pictureInPicture == pictureInPicture) {
+        return;
+    }
+
+    self._pictureInPicture = pictureInPicture;
+    if (@available(iOS 9.0, *)) {
+        if (_pipController && self._pictureInPicture && ![_pipController isPictureInPictureActive]) {
+            dispatch_async(dispatch_get_main_queue(), ^{
+                [_pipController startPictureInPicture];
+            });
+        } else if (_pipController && !self._pictureInPicture && [_pipController isPictureInPictureActive]) {
+            dispatch_async(dispatch_get_main_queue(), ^{
+                [_pipController stopPictureInPicture];
+            });
+        } else {
+            // Fallback on earlier versions
+        }
+    }
+#endif
+}
+
+- (void)setupPipController {
+  if ([AVPictureInPictureController isPictureInPictureSupported]) {
+    _pipController = [[AVPictureInPictureController alloc] initWithPlayerLayer:self._playerLayer];
+    _pipController.delegate = self;
+    // Listen if pip is available
+    // Swift code:
+    //        pipPossibleObservation = pipController.observe(\AVPictureInPictureController.isPictureInPicturePossible,
+    //options: [.initial, .new]) { [weak self] _, change in
+    //            // Update the PiP button's enabled state.
+    //            self?.pipButton.isEnabled = change.newValue ?? false
+    //        }
+  }
+}
+
+- (void)usePlayerLayer {
+    NSLog(@"Use player layer");
+    if (_player) {
+        self._playerLayer = [AVPlayerLayer playerLayerWithPlayer:_player];
+        UIViewController* vc = [[[UIApplication sharedApplication] keyWindow] rootViewController];
+        self._playerLayer.frame = CGRectMake(20, 155, 350, 350);
+        self._playerLayer.needsDisplayOnBoundsChange = YES;
+        self._playerLayer.opacity = 0.01; //TODO FIX THIS HACK -> Add layer over video player texture layer. but below the flutter layer. so we can still see the controlls
+        [vc.view.layer addSublayer:self._playerLayer];
+        vc.view.layer.needsDisplayOnBoundsChange = YES;
+        #if TARGET_OS_IOS
+            [self setupPipController];
+        #endif
+    } else {
+        NSLog(@"No player layer");
+    }
 }
 
 - (void)observeValueForKeyPath:(NSString *)path
@@ -632,6 +694,11 @@ NS_INLINE CGFloat radiansToDegrees(CGFloat radians) {
   } else {
     [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayback error:nil];
   }
+}
+
+- (void)setPictureInPicture:(FLTPictureInPictureMessage*)input error:(FlutterError**)error {
+   FLTVideoPlayer* player = self.playersByTextureId[input.textureId];
+   [player setPictureInPicture:input.enabled.intValue == 1];
 }
 
 @end
